nvda-dashboard/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
praw
yfinance
pandas
numpy
scikit-learn
torch
transformers
tensorflow
apscheduler
joblib